+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [11.1.1] - 2023-08-30
+
+### Changed
+
+- dropped jQuery
+- dropped jQuery-Lib "ColorPicker"
+- improved JavaScript performance
+- added default styling to ext_tables.sql (fixed first load without colors)
+
+### **NOTE: WORK IN PROGRESS**
 
 ## [11.0.0] - 2023-05-07